Update secrets folder structure
