+++ crates/bitwarden/src/secrets/mod.rs
@@ -1,3 +1,20 @@
-pub mod commands;
-pub mod request;
-pub mod response;
+mod create;
+mod delete;
+mod get;
+mod list;
+mod secret_response;
+mod update;
+
+pub(crate) use create::create_secret;
+pub use create::SecretCreateRequest;
+pub(crate) use delete::delete_secrets;
+pub use delete::{SecretsDeleteRequest, SecretsDeleteResponse};
+pub(crate) use get::get_secret;
+pub use get::SecretGetRequest;
+pub(crate) use list::{list_secrets, list_secrets_by_project};
+pub use list::{
+    SecretIdentifiersByProjectRequest, SecretIdentifiersRequest, SecretIdentifiersResponse,
+};
+pub use secret_response::SecretResponse;
+pub(crate) use update::update_secret;
+pub use update::SecretPutRequest;
