--- support/schemas/response/SecretDeleteResponse.json
@@ -1,54 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "Response_for_SecretDeleteResponse",
-  "type": "object",
-  "required": [
-    "success"
-  ],
-  "properties": {
-    "data": {
-      "description": "The response data. Populated if `success` is true.",
-      "anyOf": [
-        {
-          "$ref": "#/definitions/SecretDeleteResponse"
-        },
-        {
-          "type": "null"
-        }
-      ]
-    },
-    "errorMessage": {
-      "description": "A message for any error that may occur. Populated if `success` is false.",
-      "type": [
-        "string",
-        "null"
-      ]
-    },
-    "success": {
-      "description": "Whether or not the SDK request succeeded.",
-      "type": "boolean"
-    }
-  },
-  "additionalProperties": false,
-  "definitions": {
-    "SecretDeleteResponse": {
-      "type": "object",
-      "required": [
-        "id"
-      ],
-      "properties": {
-        "error": {
-          "type": [
-            "string",
-            "null"
-          ]
-        },
-        "id": {
-          "type": "string",
-          "format": "uuid"
-        }
-      },
-      "additionalProperties": false
-    }
-  }
-}
