--- support/schemas/response/SecretIdentifierResponse.json
@@ -1,57 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "Response_for_SecretIdentifierResponse",
-  "type": "object",
-  "required": [
-    "success"
-  ],
-  "properties": {
-    "data": {
-      "description": "The response data. Populated if `success` is true.",
-      "anyOf": [
-        {
-          "$ref": "#/definitions/SecretIdentifierResponse"
-        },
-        {
-          "type": "null"
-        }
-      ]
-    },
-    "errorMessage": {
-      "description": "A message for any error that may occur. Populated if `success` is false.",
-      "type": [
-        "string",
-        "null"
-      ]
-    },
-    "success": {
-      "description": "Whether or not the SDK request succeeded.",
-      "type": "boolean"
-    }
-  },
-  "additionalProperties": false,
-  "definitions": {
-    "SecretIdentifierResponse": {
-      "type": "object",
-      "required": [
-        "id",
-        "key",
-        "organizationId"
-      ],
-      "properties": {
-        "id": {
-          "type": "string",
-          "format": "uuid"
-        },
-        "key": {
-          "type": "string"
-        },
-        "organizationId": {
-          "type": "string",
-          "format": "uuid"
-        }
-      },
-      "additionalProperties": false
-    }
-  }
-}
